Separating fetch entities logic into especific hooks. #459
