+++ explorer/app/hooks/useFetchEntities.ts
@@ -1,5 +1,4 @@
-import { EntityConfig } from "app/config/common/entities";
-import { useCallback, useEffect, useMemo, useState } from "react";
+import { useCallback, useEffect, useMemo } from "react";
 import {
   AzulEntitiesResponse,
   AzulEntitiesStaticResponse,
@@ -11,15 +10,17 @@ import {
   transformTermFacets,
 } from "../apis/azul/common/filterTransformer";
 import {
+  CategoryKey,
+  CategoryValueKey,
   Pagination,
   SelectCategory,
   Sort,
-  SortOrderType,
 } from "../common/entities";
 import { useAsync } from "./useAsync";
-import { useCurrentEntity } from "./useCurrentEntity";
+import { OnFilterFn, useCategoryFilter } from "./useCategoryFilter";
 import { useFetcher } from "./useFetcher";
-import { useResetableState } from "./useResetableState";
+import { usePagination } from "./usePagination";
+import { useSort } from "./useSort";
 
 /**
  * Type of function called to update filter query string params and trigger re-fetch of entities.
@@ -32,25 +33,11 @@ export type SetFilterFn = (nextFilters: Filters) => void;
 interface EntitiesResponse {
   categories: SelectCategory[];
   loading: boolean;
+  onFilter: OnFilterFn;
   pagination?: Pagination;
   response?: AzulEntitiesResponse;
-  setFilter: SetFilterFn;
   sort?: Sort;
 }
-
-const DEFAULT_CURRENT_PAGE = 1;
-
-/**
- * Determine the default sort column.
- * @param entity - Current entity config containing all column definitions.
- * @returns Column name of default sorted column, or the first column if no default is specified.
- */
-const getDefaultSort = (entity: EntityConfig): string | undefined => {
-  return (
-    entity.list.columns.find((column) => column.sort?.default)?.sort?.sortKey ??
-    entity.list.columns[0].sort?.sortKey
-  );
-};
 
 /**
  * Hook handling the load and transformation of the values used by index pages. If the current entity loaded statically,
@@ -62,34 +49,29 @@ const getDefaultSort = (entity: EntityConfig): string | undefined => {
 export const useFetchEntities = (
   staticResponse: AzulEntitiesStaticResponse | null,
   initialFilter: Filters
-  // eslint-disable-next-line sonarjs/cognitive-complexity -- TODO revisit with #267.
 ): EntitiesResponse => {
-  // Init pagination-related state.
-  const [currentPage, setCurrentPage] = useState(DEFAULT_CURRENT_PAGE);
-
-  // Init query param-related state. TODO(mim) move to own hook to reduce complexity here, rename setFilterFn
-  const [filter, setFilter] = useState<Filters>(initialFilter);
-  const setFilterFn = useCallback((nextFilter: Filters) => {
-    setFilter(nextFilter);
-  }, []);
-
-  // Grab the current entity.
-  const entity = useCurrentEntity();
-
   // Determine type of fetch to be executed, either API endpoint or TSV.
-  const { fetchList, list, path, staticLoad } = useFetcher();
-
-  // Init sort.
-  const defaultSort = useMemo(() => getDefaultSort(entity), [entity]);
-  const [sortKey, setSortKey] = useResetableState<string | undefined>(
-    defaultSort
-  );
-  const [sortOrder, setSortOrder] = useState<SortOrderType | undefined>(
-    defaultSort ? "asc" : undefined
-  );
+  const { list, path, staticLoad } = useFetcher();
 
   // Init fetch of entities.
   const { data, isIdle, isLoading, run } = useAsync<AzulEntitiesResponse>();
+  const { resetPage, ...pagination } = usePagination(data);
+  const { sort, sortKey, sortOrder } = useSort();
+
+  // Generalize the filters returned from Azul.
+  const categories = useMemo(() => {
+    if (staticLoad || !data) {
+      return [];
+    }
+    return transformTermFacets(data.termFacets);
+  }, [data, staticLoad]);
+
+  // Init filter functionality.
+  const {
+    categories: categoryViews,
+    filter,
+    onFilter,
+  } = useCategoryFilter(categories, initialFilter);
 
   // Execute fetch of entities.
   useEffect(() => {
@@ -108,42 +90,17 @@ export const useFetchEntities = (
     }
   }, [filter, list, path, run, sortKey, sortOrder, staticLoad]);
 
-  // Handle change of sort.
-  const sort = useCallback(
-    (key?: string, order?: SortOrderType) => {
-      setSortKey(key ?? defaultSort);
-      setSortOrder(order);
+  const handleFilterChange = useCallback(
+    (
+      categoryKey: CategoryKey,
+      selectedCategoryValue: CategoryValueKey,
+      selected: boolean
+    ) => {
+      onFilter(categoryKey, selectedCategoryValue, selected);
+      resetPage();
     },
-    [defaultSort, setSortKey]
+    [onFilter, resetPage]
   );
-
-  // Create callback for next page action.
-  const nextPage = useCallback(async () => {
-    if (data?.pagination.next) {
-      setCurrentPage((s) => s + 1);
-      run(fetchList(data.pagination.next));
-    }
-  }, [data?.pagination.next, fetchList, run]);
-
-  // Create callback for previous page action.
-  const previousPage = useCallback(async () => {
-    if (data?.pagination.previous) {
-      setCurrentPage((s) => s - 1);
-      run(fetchList(data.pagination.previous));
-    }
-  }, [data?.pagination.previous, fetchList, run]);
-
-  const resetPage = useCallback(() => {
-    setCurrentPage(DEFAULT_CURRENT_PAGE);
-  }, []);
-
-  // Generalize the filters returned from Azul.
-  const categories = useMemo(() => {
-    if (staticLoad || !data) {
-      return [];
-    }
-    return transformTermFacets(data.termFacets);
-  }, [data, staticLoad]);
 
   // Exit if we're dealing with a statically-loaded entity; data has already been fetched during build; indicate
   // load is complete and return static data.
@@ -151,25 +108,18 @@ export const useFetchEntities = (
     return {
       categories: [],
       loading: false,
+      onFilter,
       response: staticResponse?.data,
-      setFilter: setFilterFn,
     };
   }
 
   // Otherwise, return the fetching, pagination and sort state.
   return {
-    categories,
+    categories: categoryViews,
     loading: isLoading || isIdle,
-    pagination: {
-      canNextPage: !!data?.pagination.next,
-      canPreviousPage: !!data?.pagination.previous,
-      currentPage,
-      nextPage,
-      previousPage,
-      resetPage,
-    },
+    onFilter: handleFilterChange,
+    pagination: { ...pagination, resetPage },
     response: data,
-    setFilter: setFilterFn,
     sort: {
       sort,
       sortKey,
